requirements.txt
awsebcli==3.14.11
blessed==1.15.0
boto3==1.9.97
botocore==1.12.97
cached-property==1.5.1
cement==2.8.2
certifi==2018.11.29
chardet==3.0.4
colorama==0.3.9
dj-database-url==0.5.0
Django==2.1.1
django-heroku==0.3.1
django-storages==1.7.1
docker==3.7.0
docker-compose==1.23.2
docker-pycreds==0.4.0
dockerpty==0.4.1
docopt==0.6.2
docutils==0.14
gunicorn==19.9.0
idna==2.7
jmespath==0.9.3
jsonschema==2.6.0
pathspec==0.5.5
Pillow==5.4.1
psycopg2==2.7.7
python-dateutil==2.8.0
pytz==2018.9
PyYAML==3.13
requests==2.20.1
s3transfer==0.2.0
semantic-version==2.5.0
six==1.11.0
termcolor==1.1.0
texttable==0.9.1
urllib3==1.24.1
wcwidth==0.1.7
websocket-client==0.54.0
whitenoise==4.1.2